Lab-7-Evaluation-Observation/requirements.txt
# Core LangChain + Snowflake Integration
langchain-snowflake>=0.2.2
langchain-core>=0.3.0
langgraph>=0.2.0

# Snowflake Connectivity
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0

# UI
streamlit>=1.28.0

# Data Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Snowflake AI Observability (TruLens)
trulens-core>=1.0.0
trulens-apps-langchain>=1.0.0
trulens-connectors-snowflake>=1.0.0
